Rename to umath to match Numpy conventions.
