+++ sparse/coo/__init__.py
@@ -1,4 +1,4 @@
 from .core import COO
-from .ufunc import elemwise
+from .umath import elemwise
 from .common import tensordot, dot, concatenate, stack, triu, tril, where, \
     nansum, nanprod, nanmin, nanmax, nanreduce
